### 2020-07-23 #2510 @whedon
[PRE REVIEW]: Bacting: a next generation, command line version of Bioclipse
**Submitting author:** @egonw (<a href="http://orcid.org/0000-0001-7542-0286">Egon Willighagen</a>)
**Repository:** <a href="https://github.com/egonw/bacting" target ="_blank">https://github.com/egonw/bacting</a>
**Version:** 0.0.11
