### 2020-07-23 #2510 @Kevin-Mattheus-Moerman
@egonw thanks for this submission. I am reviewing the suitability of this submission for JOSS and will also look for editors to handle it.
In the mean time can you please describe, for the record, the difference between the current submission and this paper to clarify this is not a duplicate publication: https://bmcbioinformatics.biomedcentral.com/articles/10.1186/1471-2105-10-397
