### 2020-07-23 #2510 @egonw
> In the mean time can you please describe, for the record, the difference between the current submission and this paper to clarify this is not a duplicate publication: https://bmcbioinformatics.biomedcentral.com/articles/10.1186/1471-2105-10-397

The Bioclipse 2 paper is still about the Eclipse-based version, not about the command line version and suffers from the limitations outlined in the "Statement of Need" section.
