### 2020-07-23 #2510 @egonw
> @egonw thanks for those suggestions. Which ones exactly are collaborators?

reality, whom I work together with in the [NanoCommons](https://github.com/nanocommons) project on ontologies.
