### 2020-07-23 #2510 @egonw
I would also not mind having someone of the original Bioclipse team look at it. They have not contributed to this adaption of it, e.g. jonalv or goglepox as they know the ins and outs of the technology, but obviously have an academic CoI.
