### 2020-07-25 #2510 @majensen
@goglepox - Would you be willing to review [this package](#2520) for the [Journal of Open Source Software](https://joss.theoj.org)? Your Bioeclipse expertise would be extremely valuable. Please have a look at [these reviewer steps](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) and let us know here whether you would have an interest. Thanks for your consideration --@majensen
