### 2020-07-25 #2510 @majensen
@Zethson - would you be willing to share your expertise in a review of [this package](#2510) for the [Journal of Open Source Software](https://joss.theoj.org) ? It is meant to improve Bioeclipse build management and command line usability, among other things. Please have a look at these [reviewer steps](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) and let us know here whether you would have an interest.
Thanks for your consideration --@majensen
