### 2020-07-26 #2510 @Zethson
Yes, I would do it. ETA 14 days is fine?  Saturday, 25 July 2020, 08:29PM +02:00 from Mark Jensen  notifications@github.com :

>@Zethson - would you be willing to share your expertise in a review of  this package for the  Journal of Open Source Software ? It is meant to improve Bioeclipse build management and command line usability, among other things. Please have a look at these  reviewer steps and let us know here whether you would have an interest.
>Thanks for your consideration -- @majensen
