### 2020-08-02 67c8ec4
Added a method to convert the report into Markdown
