### 2020-08-02 356010a
Added the xml manager, with three methods: isValid(), isWellFormed(), listNamespaces()
