### 2020-08-06 #2510 @majensen
@jonalv - Would you be willing to review [this work](#2510) for the [Journal of Open Source Software](https://joss.theoj.org)? Your expertise in Bioeclipse would be extremely valuable here. An overview of reviewing for the journal (performed openly on Github) is [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html). I appreciate your consideration very much --@majensen
