### 2020-08-07 #2510 @jonalv
Hi I can not get your link to work but from the title I see this is about
Bacting, since I know Egon Willighagen is behind a project with that name I
am hesitant as to whether I really am impartial. I have worked closely with
him, co-authoring a book with him and helped him move twice :)

// Jonathan
