### 2020-08-07 #2510 @majensen
@jonalv  - I understand, having helped @egonw move might bias your review negatively  😄
For me, your expertise is very important since it deals very directly with the application, and you have not directly contributed code to Bacting. Since you have declared possible conflict, I can take that into account.

@Kevin-Mattheus-Moerman - can you please comment here from the AIC point of view?
