### 2020-08-07 #2510 @danielskatz
If the co-authoring has been recent (last 4 years), that's another conflict.  My impression is that we should find alternative reviewers unless there is no other choice, or if @jonalv does review, then have at least 2 other reviewers.
