### 2020-08-07 #2510 @majensen
Ok @danielskatz .

Thanks for your consideration @jonalv - If you or @egonw can possibly recommend someone with Bioeclipse experience but without close collaboration, that would great. Difficult I'm sure.
