### 2020-08-07 #2510 @egonw
@majensen, reviewers do not need Bioclipse experience. The build system is a regular Maven installation. Anyone with Maven experience can build, test, and use this adaption of Bioclipse.

Please see also my original list at https://github.com/openjournals/joss-reviews/issues/2510#issuecomment-663096346 (but plz exclude Luke).
