### 2020-08-07 #2510 @majensen
@arcuri82, would you consider reviewing this work (https://github.com/openjournals/joss-reviews/issues/2510) for [the Journal of Open Source Software](https://joss.theo4j.org)? Your java expertise would be greatly appreciated. Guidelines for reviewing are [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html).
Thanks very much for considering! --@majensen
