### 2020-08-09 2e5fc1b
Added the tests for the xml manager
