### 2020-08-09 7607da1
Added mention of the preprint (submitted to JOSS)
