### 2020-08-09 b53d508
Removed old code no longer in use, as now done with 'maven release'
