### 2020-08-16 egonw/bacting#8 @Zethson
Installation tries to sign with gpg and fails
Dear @egonw,

when attempting to install with `mvn clean install` I run into:
```
[INFO] --- maven-gpg-plugin:1.5:sign (sign-artifacts) @ bacting ---
gpg: no default secret key: No secret key
gpg: signing failed: No secret key
```

The installation should not try to sign keys, or?
Am I missing something?
