### 2020-08-16 egonw/bacting#9 @Zethson
Eclipse Public License 1.0 (EPL-1.0) has been superseded
I have a question regarding the license. Why is this repository under Eclipse Public License 1.0 (EPL-1.0), when it's been superseded by Eclipse Public License 2.0 (EPL-2.0)?

Not a major issue or anything like that, but maybe you can explain that to me?

Thanks!
