### 2020-08-16 egonw/bacting#10 @Zethson
Documentation lacks a few key issues
I have a couple of suggestions regarding the documentation:

1. The README does not state in 1-2 sentences what Bacting is about. I suggest that you add 1-2 sentences from your JOSS paper so that viewers of this repository know at a glance what it is about.
2. Please mention the supported Java versions (CI pipeline gives me the answer, but still...)
3. Please state more clearly which features of the original API are not covered. I assume that the readers of the README are interested in this information and don't want to make a diff themselves.
4. I suggest that you also very quickly add a section on making releases. I doubt that any maintainers want to read the paper to find out how releases should be created.
5. Please add some community guidelines on how to contribute. I suggest a contributing.md file with a couple of sentences, maybe a code of conduct, some issue and PR templates etc. Feel free to ask me if you need clarification here.
