### 2020-08-16 #2558 @Zethson
Hi,

this is the first time that I am reviewing for JOSS, so feel free to tell me if I am doing something wrong.

1. I identified an issue with the installation instructions: https://github.com/egonw/bacting/issues/8 , which were promptly solved.
2. A question regarding the license: https://github.com/egonw/bacting/issues/9
3. Several issues regarding the community guidelines and documentation: https://github.com/egonw/bacting/issues/10

Additional questions that I have:
1. What do you, @egonw mean with missing progress monitoring? Missing progress bars for called functions?
2. How does Bioeclipse/Bacting compare to BioJava and BioPython?

The paper is well written and I am looking forward to your answers!
Great work.

Thank you!
Cheers
