### 2020-08-17 8e7e8fc
Added a short description of what Bacting is/does
