### 2020-08-17 egonw/bacting#10 @egonw
Ad 1: I added a short description of what Bacting is: https://github.com/egonw/bacting/commit/4d01cdac7d0eec6dd5637de0a1bcf7099760c911
