### 2020-08-17 #2558 @egonw
Regarding the license question, I also replied in the issue tracker, but here copy it here for convenience: There is a lot of code being reused, a good bit of which has not been written by me, but other Bioclipse developers. I cannot change the license. But I like the suggestion and will look into the (dis)advantages and discuss those with all original authors.
