### 2020-08-17 egonw/bacting#8 @egonw
I reproduced this fix on a machine without GPG key. Unless someone tells me it doesn't work, I consider this one fixed.
