### 2020-08-17 d772644
Added info which Java versions are supported.
