+++ README.md
@@ -21,7 +21,8 @@ A paper about Bacting has been writing and available from [paper/paper.md](paper
 
 # Install
 
-First, you need a working [Maven installation](https://www.google.nl/search?q=install+maven) and then install this software:
+First, you need a working [Maven installation](https://www.google.nl/search?q=install+maven) and the code is tested with
+Java 8, 11, and 14, and can be installed with:
 
 ```shell
 mvn clean install -Dgpg.skip
