### 2020-08-30 #2558 @arcuri82
I made a first set of comments.
However, there is one thing I would like to get clarified. The program is written Java, but, its intended usage seems for Groovy. I do not know Groovy. And unfortunately I do not have the time to learn a new programming language just for doing a review. If all examples are going to be only in Groovy, I am not sure I can really try out this software, and give a proper review. But, asking examples/tutorials in Java for a program that is meant to be used mainly in Groovy would likely be unfair to the author.
