### 2020-08-31 egonw/bacting#11 @egonw
Thanks for your thorough review. There is a good bit of comments that I can accommodate; thank you!

Regarding the Groovy, yes, reviewing source code in a programming language you do not used before is tricky, indeed. Any Java is basically valid Groovy, but Groovy just removes a lot of boiler plate code. But I think I can make a Java example.

Similarly, thank you for challenging me to get it working in Jython and one of the Java JavaScript implementation (as it was used in Bioclipse).

These are just some quick notes as thank you for your review. I will follow up with a full rebuttal and list of updates.
