### 2020-09-03 #2558 @egonw
Yes, I can and will. Right now, I'm swamped in university year start and project mid terms reviews. I have some ideas, some take a bit more time to work out. Other things I have to try and see how they work out. I have created a project on GitHub to work out ideas the reviewers gave, towards the rebuttal/revision: https://github.com/egonw/bacting/projects/1?add_cards_query=is%3Aopen

I will report back here when making progress. @arcuri82 wrote "first set of comments" and I will also monitor things here if more feedback comes one.
