### 2020-10-05 #2558 @kthyng
@egonw Totally understand the trouble with the start of the academic year! As the rotating associate editor in chief this week, I am going to pause this submission while we wait for your responses to start. This is for bookkeeping purposes and can be removed whenever you are ready to move forward. Thanks!
