### 2020-12-22 304ec68
Clarified TNG.
As requested in https://github.com/egonw/bacting/issues/11
