### 2020-12-22 c04f660
Clarified why the README suggesting citing the second Bioclipse paper
Per question in https://github.com/egonw/bacting/issues/11
