### 2020-12-22 2c19152
Explain what the code example is doing.
As suggested in https://github.com/egonw/bacting/issues/11
