### 2020-12-22 888e48e
Added three real world examples
