### 2020-12-23 e87de96
Added code coverage testing with JaCoCo
