+++ .travis.yml
@@ -13,3 +13,6 @@ cache:
 
 install:
   - mvn clean install -Dgpg.skip -Dmaven.javadoc.skip=true
+
+after_success:
+  - bash <(curl -s https://codecov.io/bash)
