### 2020-12-23 ec147ea
Added testing for bad input
