### 2020-12-23 3c408fb
Added the codecov badge
