### 2020-12-23 c9a7cba
Added three missing test suites
