### 2020-12-23 92e9272
Added last three test suites (tho with few methods yet)
