+++ managers-cheminfo/net.bioclipse.managers.pubchem/src/main/java/net/bioclipse/managers/PubChemManager.java
@@ -114,7 +114,7 @@ public class PubChemManager implements IBactingManager {
 
 	@Override
 	public String getManagerName() {
-		return "inchi";
+		return "pubchem";
 	}
 
 	@Override
