+++ managers-cheminfo/net.bioclipse.managers.inchi/src/test/java/net/bioclipse/managers/InChIManagerTest.java
@@ -11,6 +11,8 @@ package net.bioclipse.managers;
 
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertSame;
+import static org.junit.jupiter.api.Assertions.assertTrue;
+import static org.junit.jupiter.api.Assertions.assertFalse;
 
 import java.nio.file.Files;
 import java.util.List;
@@ -18,15 +20,25 @@ import java.util.List;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 
+import net.bioclipse.core.domain.IMolecule;
+import net.bioclipse.inchi.InChI;
+
 public class InChIManagerTest {
 
 	static InChIManager inchi;
+	static CDKManager cdk;
 	static String workspaceRoot;
 
 	@BeforeAll
 	static void setupManager() throws Exception {
 		workspaceRoot = Files.createTempDirectory("inchitestws").toString();
 		inchi = new InChIManager(workspaceRoot);
+		cdk = new CDKManager(workspaceRoot);
+	}
+
+	@Test
+	public void testManagerName() {
+		assertSame("inchi", inchi.getManagerName());
 	}
 
 	@Test
@@ -35,4 +47,25 @@ public class InChIManagerTest {
 		assertNotNull(dois);
 		assertSame(0, dois.size());
 	}
+
+	@Test
+	public void testGenerate() throws Exception {
+		IMolecule mol = cdk.fromSMILES("CC");
+		InChI someInChI = inchi.generate(mol);
+		assertNotNull(someInChI);
+		assertTrue(someInChI.getValue().contains("InChI="));
+	}
+
+	@Test
+	public void testCheck() throws Exception {
+		assertFalse(inchi.check("InChI="));
+		assertTrue(inchi.check("InChI=1S/CH4/h1H4"));
+	}
+	
+	@Test
+	public void testCheckKey() throws Exception {
+		assertFalse(inchi.checkKey("VNWKTOKETHGBQD-UHFFFAOYSA-3"));
+		assertTrue(inchi.checkKey("VNWKTOKETHGBQD-UHFFFAOYSA-N"));
+	}
+
 }
