### 2020-12-24 74151e3
More testing for the CDK manager
