### 2020-12-24 7e1c30e
Added instructions on how releases are made
