### 2020-12-26 a2e9356
Added two methods used to generate non-Standard InChIs
