### 2020-12-26 egonw/bacting#39 @egonw
BiojavaManager
- [ ] public List<IDNA> DNAsFromString(String dnaString) {
 - [ ] public List<IRNA> RNAsFromString(String rnaString) {
 - [ ] public List<IProtein> proteinsFromString(String proteinString) {
 - [ ] public IDNA DNAfromPlainSequence(String dnaString) {
 - [ ] public IRNA RNAfromPlainSequence(String rnaString) {
 - [ ] public IProtein proteinFromPlainSequence(String proteinString) {
 - [ ] public IDNA DNAfromPlainSequence(String dnaString, String name) {
 - [ ] public IRNA RNAfromPlainSequence(String rnaString, String name) {
 - [ ] public IProtein proteinFromPlainSequence(String proteinString,
 - [ ] public IProtein DNAtoProtein(IDNA dna) {
 - [ ] public IProtein DNAtoProtein(IDNA dna, String name) {
 - [ ] public List<IProtein> DNAtoProtein(List<IDNA> dnas) {
 - [ ] public IRNA DNAtoRNA(IDNA dna) {
 - [ ] public IRNA DNAtoRNA(IDNA dna, String name) {
 - [ ] public List<IRNA> DNAtoRNA(List<IDNA> dnas) {
 - [ ] public IProtein RNAtoProtein(IRNA rna) {
 - [ ] public IProtein RNAtoProtein(IRNA rna, String name) {
 - [ ] public List<IProtein> RNAtoProtein(List<IRNA> rnas) {
 - [ ] public IRNA transcriptionOf(IDNA dna) {
 - [ ] public IRNA transcriptionOf(IDNA dna, String name) {
 - [ ] public List<IRNA> transcriptionOf(List<IDNA> dnas) {
 - [ ] public IProtein translationOf(IRNA rna) {
 - [ ] public IProtein translationOf(IRNA rna, String name) {
 - [ ] public List<IProtein> translationOf(List<IRNA> rnas) {
 - [ ] public String getManagerName() {
 - [ ] public List<IDNA> DNAsFromFile(IFile file)
 - [ ] public List<IRNA> RNAsFromFile(IFile file)
 - [ ] public List<IProtein> proteinsFromFile(IFile file)
 - [ ] public List<ISequence> sequencesFromFile( IFile file )
 - [ ] public void sequencesToFASTAfile(final List<? extends ISequence> sequences,
 - [ ] public String sequencesToFASTAString(
 - [ ] public RecordableList<ISequence> createSequenceList()
 - [ ] public String proteinsToFASTAString(final List<IProtein> proteins){
 - [ ] public String dnaToFASTAString(final List<IDNA> dnas){
 - [ ] public void proteinsToFASTAfile(List<IProtein> proteins, IFile file) {
 - [ ] public List<IFeature> features(ISequence sequence) {
 - [ ] public List<ISequence> sequencesFromString(String s) {
