### 2020-12-29 7eccc98
Make sure the data sources are initialized
