### 2020-12-29 edf928a
Define the bio data sources earlier
