### 2020-12-29 ee473cd
Replaced outdated APIs with new methods
