+++ .travis.yml
@@ -1,18 +1,0 @@
-language: java
-
-jdk:
-- openjdk8
-- openjdk11
-- openjdk14
-
-sudo: false
-
-cache:
-  directories:
-  - $HOME/.m2
-
-install:
-  - mvn clean install -Dgpg.skip -Dmaven.javadoc.skip=true
-
-after_success:
-  - bash <(curl -s https://codecov.io/bash)
