### 2020-12-29 9b0aa4b
More testing for the inchi manager
