+++ .github/workflows/maven.yml
@@ -22,3 +22,5 @@ jobs:
         java-version: ${{ matrix.java }}
     - name: Build with Maven
       run: mvn clean test -q
+    - name: push JaCoCo stats to codecov.io
+      run: bash <(curl -s https://codecov.io/bash)
