+++ .github/workflows/maven.yml
@@ -21,6 +21,6 @@ jobs:
       with:
         java-version: ${{ matrix.java }}
     - name: Build with Maven
-      run: mvn clean test -q
+      run: mvn clean install -Dgpg.skip -Dmaven.javadoc.skip=true
     - name: push JaCoCo stats to codecov.io
       run: bash <(curl -s https://codecov.io/bash)
