### 2020-12-29 ca1843d
More testing of the RDFManager
