### 2020-12-29 6e8dcfa
Also test parsing as CML string
