### 2020-12-29 9a1113b
Fixed the test to test the correct method
