### 2020-12-29 81cdda9
More testing in bioclipse and ui
