### 2020-12-29 6e8bfa2
Initial bit of biojava manager functionality
