### 2020-12-29 6725ff1
Added pubchem.downloadAsString()
