### 2020-12-30 92a7b1a
A few corner case tests to boost the code coverage
