### 2020-12-30 115f05c
Testing the ISequence a bit more
