+++ managers-bioinfo/net.bioclipse.managers.biojava/src/test/java/net/bioclipse/managers/BiojavaManagerTest.java
@@ -12,6 +12,7 @@ package net.bioclipse.managers;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertSame;
 import static org.junit.jupiter.api.Assertions.assertThrows;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import java.io.IOException;
 import java.nio.file.Files;
@@ -49,6 +50,11 @@ public class BiojavaManagerTest {
 		assertNotNull(dna.getPlainSequence());
 		dna = biojava.DNAfromPlainSequence("CAT", "foo");
 		assertSame("foo", dna.getName());
+		String fasta = dna.toFasta();
+		assertNotNull(fasta);
+		assertTrue(fasta.startsWith(">"));
+		Object parsedResource = dna.getParsedResource();
+		assertNotNull(parsedResource);
 	}
 
 	@Test
@@ -69,6 +75,11 @@ public class BiojavaManagerTest {
 		assertNotNull(protein.toString());
 		assertNotNull(protein.getPlainSequence());
 		assertNotNull(protein.getName());
+		String fasta = protein.toFasta();
+		assertNotNull(fasta);
+		assertTrue(fasta.startsWith(">"));
+		Object parsedResource = protein.getParsedResource();
+		assertNotNull(parsedResource);
 	}
 
 }
