### 2020-12-30 f59019d
Fixed BiojavaSequence.toFasta() which did not use the right approach to convert a byte[] to a String
