+++ managers-bioinfo/net.bioclipse.managers.biojava/src/main/java/net/bioclipse/biojava/domain/BiojavaSequence.java
@@ -64,7 +64,7 @@ public class BiojavaSequence extends BioObject implements ISequence {
             throw new IllegalStateException("Illegal BiojavaSequence", e);
         }
 
-        return os.toByteArray().toString();
+        return new String(os.toByteArray());
     }
 
     /**
