### 2020-12-30 c389f08
Test the Biojava constructors
