### 2020-12-30 c62ecf2
Updated the build badge
