### 2020-12-30 c0b60bd
Some example from 'A lot of Bioclipse Scripting Language examples'
