### 2020-12-30 08b8513
Updated the paper for some recent updates
- new biojava manager
- JaCoCo/codecov.io test coverage
- newer OPSIN and BridgeDb releases
- replaced Travis CI with GitHub Actions
- be more clear about OSGi and Spring
