+++ paper/paper.bib
@@ -1,3 +1,7 @@
+@article{Holland2008,
+  title={BioJava: an open-source framework for bioinformatics}, volume={24}, ISSN={1460-2059}, DOI={10.1093/bioinformatics/btn397}, abstractNote={Summary: BioJava is a mature open-source project that provides a framework for processing of biological data. BioJava contains powerful analysis and statistical routines, tools for parsing common file formats and packages for manipulating sequences and 3D structures. It enables rapid bioinformatics application development in the Java programming language.}, number={18}, journal={Bioinformatics}, author={Holland, R. C. G. and Down, T. A. and Pocock, M. and Prlić, A. and Huen, D. and James, K. and Foisy, S. and Dräger, A. and Yates, A. and Heuer, M. and et al.}, year={2008}, month={Sep}, pages={2096–2097}
+}
+
 @article{zenodo,
 	Abstractnote = {<p>Gala is a Python package for Galactic astronomy and gravitational dynamics. The bulk of the package centers around implementations of gravitational potentials, numerical integration, and nonlinear dynamics.</p>},
 	Author = {Egon Willighagen},
