+++ .github/ISSUE_TEMPLATE/feature-request--bioclipse-api-method.md
@@ -1,0 +1,20 @@
+---
+name: 'Feature request: Bioclipse API method'
+about: Suggest a Bioclipse API method to be ported
+title: ''
+labels: enhancement
+assignees: ''
+
+---
+
+## Bioclipse API method to be ported
+
+A short description of the Bioclipse API method you like to see ported. Ideally, provide as much context as possible, such as journal articles describing or using the manager, links to source code, etc.
+
+## Example use
+
+A short code example where the API method is used.
+
+## Additional context
+
+Add any other context or screenshots about the feature request here.
