### 2020-12-31 f3f6551
Added an initial contributors documentation.
