+++ .github/ISSUE_TEMPLATE/bug-report--unexpected-output-of-bacting-api-method.md
@@ -1,0 +1,28 @@
+---
+name: 'Bug report: unexpected output of Bacting API method'
+about: Create a report to help us improve
+title: ''
+labels: bug
+assignees: ''
+
+---
+
+## Bacting API method with unexpected output
+
+A short description of the unexpected output. Please include the code you used, and the output in Bacting and what you expected or got with Bioclipse below:
+
+### Expected Output
+
+```
+The expected output
+```
+
+### Actual Output
+
+```
+The actual output with Bacting
+```
+
+## Additional context
+
+Add any other context or screenshots about the feature request here.
