+++ .github/PULL_REQUEST_TEMPLATE/pull_request_template.md
@@ -1,0 +1,11 @@
+# Patch description
+
+A concise description of what this patch does. Where appropriate, mention issue reports, etc.
+
+# Pull request checks
+
+When reusing Bioclipse code, please ensure:
+
+ - [ ] does all code have proper copyright and license info?
+ - [ ] all reused code has an OSI-approved license?
+ - [ ] all new code has an Eclipse Public License (1.0 or later and GPL exception)?
