+++ .github/PULL_REQUEST_TEMPLATE/pull_request_template.md
@@ -2,6 +2,12 @@
 
 A concise description of what this patch does. Where appropriate, mention issue reports, etc.
 
+Changes proposed in this pull request:
+
+ - 
+ - 
+ - 
+
 # Pull request checks
 
 When reusing Bioclipse code, please ensure:
@@ -9,3 +15,5 @@ When reusing Bioclipse code, please ensure:
  - [ ] does all code have proper copyright and license info?
  - [ ] all reused code has an OSI-approved license?
  - [ ] all new code has an Eclipse Public License (1.0 or later and GPL exception)?
+
+@egonw
