### 2020-12-31 #2558 @egonw
> 2. A question regarding the license: [egonw/bacting#9](https://github.com/egonw/bacting/issues/9)

This one is still open. There are several past Bioclipse developers involved, one of them even passed away. I will not be able to resolve this quickly.

> 3. Several issues regarding the community guidelines and documentation: [egonw/bacting#10](https://github.com/egonw/bacting/issues/10)

@Zethson, I have isolated some subtasks of https://github.com/egonw/bacting/issues/10 which now have a reply (fixes mostly; see that issue). Thank you for your suggestions.

> Additional questions that I have:
>
> 1. What do you, @egonw mean with missing progress monitoring? Missing progress bars for called functions?

Indeed. Eclipse (and therefore Bioclipse) has a progress monitoring `View` (a GUI element) that allows you to see which processes are running in the background. For long running tasks, it would tell the user what they are waiting for. But since Bacting does not have a GUI, I have not ported this functionality yet. I have been thinking about this, but it depends a lot on how it is used. I cannot exclude I will include the progress monitoring in the future, but at this moment this does not have priority.

I'll update the manuscript to be more clear about this.

> 2. How does Bioeclipse/Bacting compare to BioJava and BioPython?

BioJava is a library that actually has been integrated into (the original) Bioclipse. This week I [started](https://github.com/egonw/bacting/issues/39) [porting](https://bioclipse.github.io/bioclipse.scripting/bio.html#the-biojava-manager) the functionality. Bioclipse exposes the functionality and makes it interoperable with other libraries by introducing implementation-neutral 'domain objects'.
