### 2020-12-31 #2558 @egonw
> @egonw - checking in. How are things going?

@majensen, yes, sorry about this. i had to wait until the winter break to start working on the rebuttal. I think I indicated somewhere that I expect to "resubmit" (how does that work?) the manuscript before this Monday. I am tracking progress here: https://github.com/egonw/bacting/projects/1
