### 2020-12-31 egonw/bacting#43 @egonw
check if paper can be updated for progress monitoring
> @egonw mean with missing progress monitoring

See: https://github.com/openjournals/joss-reviews/issues/2558
