### 2020-12-31 c8dbd7c
Updated the copyright section, as it still reflected a very old version of this repository
