+++ README.md
@@ -115,7 +115,7 @@ modules cached by Groovy, by doing something like, to remove earlier SNAPSHOT ve
 
 # Copyright and authors
 
-All code in the `/bioclipse/` folder contains Bioclipse 2 code and the headers of the individual
-source code files describe who contributed to that code of that class, but unfortunately this code
+Code in this repository contains mostly code that originated from Bioclipse
+and the headers of the individual source code files describe who contributed to that code of that class, but unfortunately this code
 ownership is not always clear. I refer to the various [Bioclipse code repositories](https://github.com/bioclipse)
-for the git history for more information.
+for the git history for detailed information.
