### 2020-12-31 6a4c007
Added a link to Bioclipse
Link to the Github organization as the original bioclipse.net website is no longer.
