### 2021-01-08 3360ca5
Ported cdk.isValid(casNumber)
