### 2021-03-31 6943d23
Ported the cdk.partition() method
