### 2021-04-03 3138894
Some clarifications on installing Bacting
