### 2021-04-25 egonw/bacting#10 @egonw
I think I have addressed all these points now.
