### 2021-04-25 egonw/bacting#11 @egonw
@arcuri82, you asked "Why there is the section for "Grabbing Bacting from Groovy"?". I have this section in here to show the difference with the original Bioclipse script. I prefer to keep it.
