### 2021-04-25 egonw/bacting#11 @egonw
I think I got all issues, except the compile questions, which is an open issue (at the time of writing): https://github.com/egonw/bacting/issues/23
