### 2021-04-25 #2558 @egonw
One of the last hurdles I wanted to take before the resubmission is using it in Python. There is also @arcuri82's question about Java. Since the manager is written in Java, you can simply use them directly in Java too. The dependencies which are resolved in Groovy with `@Grab` and in Python with `scyjava.config` (see the updated 0.17 README) can be defined in (for example) Maven. And creating a manager, like in the `cdk.fromSMILES("COC")` is done with the equivalent of:

```java
import net.bioclipse.managers.CDKManager;
CDKManager cdk = new CDKManager(".");
```

Of course, it needs the `main()` wrapping, which is why I prefer Groovy, that does not.

**Rebuttal**

First, thanks to the reviewers for their detailed review. It is my first JOSS submission and I really appreciate the thoroughness, beating the peer review we once established for the Chemistry Development Kit.

Since the issues got scattered over various comments in this issue and in the Bacting GitHub issue, I actually went through them and extracted several items for which I created issues: https://github.com/egonw/bacting/projects/1 You can find the detailed rebuttal there.

Roughly, code coverage added (70%), more detailed guidance on GitHub (like contributing.md), replied the questions about unclear things, or why something was done, added missing JavaDoc, and added a tracker for the porting of the Bioclipse API (see https://github.com/egonw/bacting/projects/2).

As said, I haven't been able to reproduce the compile issue, but solved the JavaDoc error in the report. That leaves two open questions, I believe:

1. the EPL version. contacting all authors of the original code is going to take some time. I suggest to decouple this from the paper.
2. the progress monitoring. This has not been ported from Bioclipse to Bacting at this moment. Eclipse has a nice progress user interface, but reproducing this on the command line is non-trivial. Maybe it will return at some point, but I do not foresee this in the near future.

@majensen, please consider this my point to continue the review process. I am sorry it took so long. I clearly underestimated to level of detail the reviewers would give to the work and generally fully agree with their comments. It greatly improved the project.
