### 2021-04-26 46f7ee2
Added new bridgedb.register(code, name)
