### 2021-04-27 f5ae71f
Ported the OscarManager
