### 2021-04-27 bd9a675
Added the third method
