### 2021-06-01 #2558 @egonw
@majensen, I got an editorial question: for another journal we're piloting using CiTO annotations in reference lists, see https://www.biomedcentral.com/collections/cito If I can get this to work with the current template and JOSS compiler out of the box, would JOSS be okay with me adding some of them?
