### 2021-06-03 #2558 @majensen
@egonw I will tag @arfon regarding the [above](https://github.com/openjournals/joss-reviews/issues/2558#issuecomment-852336355) proposal anyhow.
