### 2021-06-14 #2558 @majensen
@egonw  - I have made some very minor suggestions to smooth out the English:
[10.21105.joss.02558-maj.pdf](https://github.com/openjournals/joss-reviews/files/6645339/10.21105.joss.02558-maj.pdf). While I'm not in the chem business, I can certainly appreciate the new life Bacting appears to be breathing into a useful and venerable toolkit.

Once you have made these changes, can you provide me with a Zenodo archive DOI of the repo here. This may need to be an ad hoc archive, as JOSS prefers the archive to have the precise title and author list as the paper it refers to. Thank you!
