### 2021-06-14 #2558 @egonw
@majensen, just to make sure... a new Zenodo archive, just for the paper, right? You do not mean to update the paper for the most recent Bacting release, right?
