### 2021-06-14 a195cac
Applied changes by JOSS editor
See https://github.com/openjournals/joss-reviews/issues/2558#issuecomment-860332546
