### 2021-06-14 #2558 @egonw
The Bacting version described in the article is 0.15, with DOI: https://doi.org/10.5281/zenodo.4400601

The DOI for the updated Bacting paper: to follow asap, awaiting PDF generation by Whedon...
