### 2021-06-21 #2558 @majensen
@egonw - for the archive, can you rename your latest to the paper title, viz. "Bacting: a next generation, command line version of Bioclipse" ? Then I can use that with the version "paper-V3", or you might call it 0.18.1, or what have you.
