### 2021-06-22 #2558 @majensen
Sorry, I'm not being clear @egonw  - so, JOSS likes the title _of the archive_ to match the title _of the paper_. Right now, the title of the archive at https://doi.org/10.5281/zenodo.4942022 is "egon/bacting: JOSS paper v3".
So, in the Zenodo UI, where it says "title" in the Basic Information form (see screenshot below), one would put "Bacting: a next generation...".
Does this make sense?
