### 2021-06-22 #2558 @egonw
Done: https://zenodo.org/record/4942022
