### 2021-06-22 #2558 @kyleniemeyer
Hi @egonw, I'm doing some final checks before publishing. The Brenninkmeijer, Mayfield, Rijswijk, Slenter, Spjuth 2009, and Willighagen 2017 references look a bit strange - I think "et al." is included in the bibtex entry somehow. Can you correct those? All authors should be included in the bibtex entries, and then our LaTeX parser can handle truncating the list of authors automatically.
