+++ paper/paper.bib
@@ -53,11 +53,23 @@
 }
 
 @article{Mayfield2019,
-  title={cdk/cdk: CDK 2.3}, url={https://zenodo.org/record/3364510}, DOI={10.5281/ZENODO.3364510}, abstractNote={The Chemistry Development Kit}, publisher={Zenodo}, author={Mayfield, John and Willighagen, Egon and Rajarshi Guha and Gilleain Torrance and Ujihara, Kazuya and Rahman, Syed Asad and Alvarsson, Jonathan and Vine, Mark B and Gražulis, Saulius and Tomáš Pluskal and et al.}, year={2019}, month={Aug}
+  title={cdk/cdk: CDK 2.3},
+  url={https://zenodo.org/record/3364510},
+  DOI={10.5281/ZENODO.3364510},
+  publisher={Zenodo},
+  author={Mayfield, John and Willighagen, Egon and Rajarshi Guha and Gilleain Torrance and Ujihara, Kazuya and Rahman, Syed Asad and Alvarsson, Jonathan and Vine, Mark B and Gražulis, Saulius and Pluskal, Tomáš and Wei, Yap Chun and Szisz, Daniel and Williamson, Mark J. and Kochev, Nikolay and Jeliazkova, Nina and Bach, Eric and Berg, Arvid and Clark, Alex and Stephan, Ralf and Wenk, Michael and ficolas2 and Stueker, Oliver and Dole, Krishna and Jönsson, Klas and kaibioinfo and Burgoon, Lyle and Katsubo, Dmitry and AndyHowlettGitHub and Köhler, Uli and Harmon, Cyrus},
+  year={2019},
+  month={Aug}
 }
 
 @article{Brenninkmeijer2020,
-  title={bridgedb/BridgeDb: BridgeDb 2.3.8}, url={https://zenodo.org/record/3855084}, DOI={10.5281/ZENODO.3855084}, abstractNote={This minor release adds Guide to Pharmacology Targets as identifier sources. The below examples assume you need just <code>org.bridgedb.bio</code>, and you can use it as template and modify to add other modules. For Maven: <pre><code class="lang-xml">&lt;dependencies&gt; &lt;dependency&gt; &lt;groupId&gt;org.bridgedb&lt;/groupId&gt; &lt;artifactId&gt;org.bridgedb.bio&lt;/artifactId&gt; &lt;version&gt;2.3.8&lt;/version&gt; &lt;/dependency&gt; &lt;/dependencies&gt; </code></pre> For Gradle: <pre><code class="lang-gradle">compile group: “org.bridgedb”, name: “org.bridgedb.bio”, version: “2.3.8” </code></pre> For Ivy: <pre><code class="lang-xml">&lt;dependency org="org.bridgedb" name="org.bridgedb.bio" rev="2.3.8"/&gt; </code></pre> For Buildr: <pre><code class="lang-buildr">’org.bridgedb:org.bridgedb.bio:jar:2.3.8’ </code></pre>}, publisher={Zenodo}, author={Brenninkmeijer, Christian Y. and Willighagen, Egon and Manas Awasthi and Soiland-Reyes, Stian and Riutta, Anders and Pico, Alexander and Dunlop, Ian and Nunogit and Kerber, Randy and De and et al.}, year={2020}, month={May}
+  title={bridgedb/BridgeDb: BridgeDb 2.3.8},
+  url={https://zenodo.org/record/3855084},
+  DOI={10.5281/ZENODO.3855084},
+  publisher={Zenodo},
[…]
