### 2021-06-23 #2558 @egonw
@kyleniemeyer, thank you for catching that. I had not expected Zotero to put et al. in their BibTeX output :( There was a second entry to which this applied and fixed that one too.

When looking at the final proof last night I noticed one sentence that changed in reply to a reviewer comment, but where something seems to have gone wrong. I am changing `This means that translation of IFile to String translations in the API do not exist in Bioclipse` to `This means that the IFile to String translations in the API do not exist in Bioclipse.`, see https://github.com/egonw/bacting/commit/ff7414cd3990639019e3a80d3e36feb0b95791bf

- [ ] make new proof and check changes
- [ ] make a new tag `paper-v4` for the article in the repo (for the `whedon set paper-v4 as version`)
- [ ] wait for the Zenodo deposit and look up the new DOI on Zenodo (for the `whedon set 10.5281/zenodo.494xxxx as archive)
