### 2021-06-23 #2558 @kyleniemeyer
@egonw ah sorry for the confusion - I actually meant all *six* references that began with those first authors...

You do not actually need to update the Zenodo deposit for paper-only changes—you can, if you want, but our focus is on archiving the version of the software that was reviewed, since JOSS archives the paper itself.
