### 2021-06-23 faaa5da
The Eclipse core dependency no longer supports Java 8
