+++ .github/workflows/maven.yml
@@ -11,7 +11,7 @@ jobs:
     runs-on: ubuntu-latest
     strategy:
       matrix:
-        java: [ 8, 11, 14 ]
+        java: [ 11, 14 ]
     name: Java ${{ matrix.java }}
 
     steps:
