### 2021-06-23 #2558 @kyleniemeyer
Actually it looks like all of the references with long author lists have this issue.
