+++ paper/paper.bib
@@ -1,86 +1,269 @@
 @article{Holland2008,
-  title={BioJava: an open-source framework for bioinformatics}, volume={24}, ISSN={1460-2059}, DOI={10.1093/bioinformatics/btn397}, abstractNote={Summary: BioJava is a mature open-source project that provides a framework for processing of biological data. BioJava contains powerful analysis and statistical routines, tools for parsing common file formats and packages for manipulating sequences and 3D structures. It enables rapid bioinformatics application development in the Java programming language.}, number={18}, journal={Bioinformatics}, author={Holland, R. C. G. and Down, T. A. and Pocock, M. and Prlić, A. and Huen, D. and James, K. and Foisy, S. and Dräger, A. and Yates, A. and Heuer, M. and et al.}, year={2008}, month={Sep}, pages={2096–2097}
+	title = {{BioJava}: an open-source framework for bioinformatics},
+	volume = {24},
+	issn = {1460-2059},
+	url = {http://dx.doi.org/10.1093/bioinformatics/btn397},
+	doi = {10.1093/bioinformatics/btn397},
+	abstract = {Summary: BioJava is a mature open-source project that provides a framework for processing of biological data. BioJava contains powerful analysis and statistical routines, tools for parsing common file formats and packages for manipulating sequences and 3D structures. It enables rapid bioinformatics application development in the Java programming language.},
+	number = {18},
+	journal = {Bioinformatics},
+	author = {Holland, R. C. G. and Down, T. A. and Pocock, M. and Prlić, A. and Huen, D. and James, K. and Foisy, S. and Dräger, A. and Yates, A. and Heuer, M. and Schreiber, M. J.},
+	month = sep,
+	year = {2008},
+	pmid = {18689808},
+	keywords = {bioinformatics, java},
+	pages = {2096--2097}
 }
 
 @article{zenodo,
-	Abstractnote = {<p>Gala is a Python package for Galactic astronomy and gravitational dynamics. The bulk of the package centers around implementations of gravitational potentials, numerical integration, and nonlinear dynamics.</p>},
-	Author = {Egon Willighagen},
-	Doi = {10.5281/zenodo.3950863},
+	Author = {Willighagen, Egon},
+	Doi = {10.5281/zenodo.4400601},
 	Month = {Jul},
 	Publisher = {Zenodo},
-	Title = {egonw/bacting: Bacting 0.0.11},
+	Title = {{egonw/bacting: Bacting 0.0.15}},
 	Year = {2020}
 }
 
 @article{bioclipse1,
-  title={Bioclipse: an open source workbench for chemo- and bioinformatics}, volume={8}, ISSN={1471-2105}, DOI={10.1186/1471-2105-8-59}, abstractNote={There is a need for software applications that provide users with a complete and extensible toolkit for chemo- and bioinformatics accessible from a single workbench. Commercial packages are expensive and closed source, hence they do not allow end users to modify algorithms and add custom functionality. Existing open source projects are more focused on providing a framework for integrating existing, separately installed bioinformatics packages, rather than providing user-friendly interfaces. No open source chemoinformatics workbench has previously been published, and no sucessful attempts have been made to integrate chemo- and bioinformatics into a single framework.}, number={1}, journal={BMC Bioinformatics}, author={Spjuth, Ola and Helmus, Tobias and Willighagen, Egon L. and Kuhn, Stefan and Eklund, Martin and Wagener, Johannes and Murray-Rust, Peter and Steinbeck, Christoph and Wikberg, Jarl E.}, year={2007}, month={Feb}, pages={59+}
[…]
