### 2021-06-23 #2558 @egonw
Okay, took a bit of effort, but I exported all references in a different way from Zotero now (next time I'm just going to use `citation.js`)... I think it should be better now, but also note the following, which comes from the PDF generation process, I think:

![image](https://user-images.githubusercontent.com/26721/123047070-21ab6780-d3fd-11eb-8444-79dd08b07ccc.png)
