### 2021-06-23 #2558 @egonw
> You do not actually need to update the Zenodo deposit for paper-only changes—you can, if you want, but our focus is on archiving the version of the software that was reviewed, since JOSS archives the paper itself.

Okay, that seems to me to contradict earlier instructions, but will do without now then. @kyleniemeyer, please do check the references one more time too.
