### 2021-06-23 3937fbf
One of the dependencies no longer support Java 8
