### 2021-06-23 #2558 @kyleniemeyer
@egonw Thanks for your work on that, it now looks good to me. Our paper template does truncate some very long author lists, but includes the final author.
