### 2021-06-23 #2558 @egonw
> Congratulations @egonw on your article's publication in JOSS!
>
> Many thanks to @Zethson and @arcuri82 for reviewing, and @majensen for editing.

Thanks everyone for landing this work!
